- Some work on Warehouse class
-Edited the AddStock method in Warehouse
-Edited the WarehouseRent property
-Added hard coded warehouseRent in the constructor
-Added new method FilledVolume

-Edited the form for Warehouse

-Other minor work

+++ Supermarket.Models/Interfaces/IStock.cs
@@ -4,7 +4,7 @@
     {
         //Everything here explains itself. No need for additional explanation.
         string ProductName { get; }
-        decimal Quantity { get; }
+        int Quantity { get;}
         decimal Price { get; }
         int DaysTillExpiration { get; }
         void ChangePrice(decimal newPrice);
